Fixes #14 and #15
For #14: I've added the proper SQS import and setup code
For #15: I've added a check against the error code. If the code is 21211 (twilio invalid phone code) then we will log out at INFO level "twilio rejected mobile as invalid - sms not sent for verification code 12345 and will not be retried" (with the appropriate verification code.